today update

+++ controller/rewardController.js
@@ -1,33 +1,34 @@
+import mongoose from "mongoose"
 import { reward } from "../model/rewardModel.js"
 import { ErrorHandler } from "../utils/Errorhandler.js"
 
-export const createReward = async(req, res, next)=>{
+export const createReward = async (req, res, next) => {
     try {
 
         const {
-                type, 
-                value, 
-                probability,
-                isActive, 
-                expiresAt 
+            type,
+            value,
+            probability,
+            isActive,
+            expiresAt
         } = req.body || {}
 
 
 
         const data = await reward.create({
-                     //"coins", "cashback", "voucher", "free_delivery", "none"
-            type, 
-            value, 
+            //"coins", "cashback", "voucher", "free_delivery", "none"
+            type,
+            value,
             probability,
             isActive,
             expiresAt
         })
 
         res.status(200).json({
-            success : true,
+            success: true,
             data
         })
-        
+
     } catch (err) {
         console.error(err)
         return next(new ErrorHandler(`${err._message}`, 500))
@@ -35,17 +36,17 @@ export const createReward = async(req, res, next)=>{
 }
 
 
-export const getRewardList = async(req, res, next)=>{
+export const getRewardList = async (req, res, next) => {
     try {
 
         const data = await reward.find({})
 
-       
+
         res.status(200).json({
-            success : true,
+            success: true,
             data
         })
-        
+
     } catch (err) {
         console.error(err)
         return next(new ErrorHandler(`${err._message}`, 500))
@@ -53,38 +54,43 @@ export const getRewardList = async(req, res, next)=>{
 }
 
 
-export const editReward = async(req, res, next)=>{
+export const editReward = async (req, res, next) => {
     try {
 
         const {
-                type, 
-                value, 
-                probability,
-                isActive, 
-                expiresAt 
+            type,
+            value,
+            probability,
+            isActive,
+            expiresAt
         } = req.body || {}
 
         const rewardId = req.params.id
 
+        if (!mongoose.Types.ObjectId.isValid(rewardId)) {
+            return next(new ErrorHandler("Invalid reward ID", 400));
+        }
 
-        const data = await reward.findOne({_id : rewardId})
+        const data = await reward.findOne({ _id: rewardId })
 
-            return next(new ErrorHandler("reward id is invaild", 400))
-        if(type) data.type = type
-        if(value) data.value = value
-        if(probability) data.probability = probability
-        if(typeof isActive === "boolean") data.isActive = isActive
-        if(expiresAt) data.expiresAt = expiresAt
+        if (!data) {
+            return next(new ErrorHandler("Reward not found", 404));
+        }
+        if (type) data.type = type
+        if (value) data.value = value
+        if (probability) data.probability = probability
+        if (typeof isActive === "boolean") data.isActive = isActive
+        if (expiresAt) data.expiresAt = expiresAt
 
 
 
         await data.save()
 
         res.status(200).json({
-            success : true,
+            success: true,
             data
         })
-        
+
     } catch (err) {
         console.error(err)
         return next(new ErrorHandler(`${err._message}`, 500))
@@ -93,21 +99,26 @@ export const editReward = async(req, res, next)=>{
 
 
 
-export const deleteReward = async(req, res, next)=>{
+export const deleteReward = async (req, res, next) => {
     try {
         const rewardId = req.params.id
+          if (!mongoose.Types.ObjectId.isValid(rewardId)) {
+            return next(new ErrorHandler("Invalid reward ID", 400));
+        }
 
-        const data = await reward.findOne({_id : rewardId})
+        const data = await reward.findOne({ _id: rewardId })
 
-            return next(new ErrorHandler("reward id is invaild", 400))
-        
+
+        if (!data) {
+            return next(new ErrorHandler("Reward not found", 404));
+        }
         await data.deleteOne()
 
         res.status(200).json({
-            success : true,
-            message : "reward deleted successfully !"
+            success: true,
+            message: "reward deleted successfully !"
         })
-        
+
     } catch (err) {
         console.error(err)
         return next(new ErrorHandler(`${err._message}`, 500))
@@ -115,30 +126,33 @@ export const deleteReward = async(req, res, next)=>{
 }
 
 
-export const randomreward = async(req, res, next)=>{
-     try {
-       const data = await reward.find({isActive : true})
+export const randomreward = async (req, res, next) => {
+    try {
+        const data = await reward.find({ isActive: true })
 
-       const total = data.reduce((sum, r)=> sum + r.probability, 0)
-       
-    //    console.log(total)
+        const total = data.reduce((sum, r) => sum + r.probability, 0)
 
-       let random = Math.random() * total;
+           console.log(total)
 
-       let selectreward ;
+        let random = Math.random() * total;
 
-       for(const reward of data){
-        if(random < reward.probability){
-            selectreward = reward
-            break;
+        let selectreward = null;
+
+        for (const reward of data) {
+            if (random < reward.probability) {
+                selectreward = reward
+                break;
+            }
+            random -= reward.probability
         }
-        random -= reward.probability
-       }
 
-        return null
-        
-       res.send(selectreward)        
-      
+        // return null
+
+       return res.status(200).json({
+            success: true,
+            reward: selectreward
+        });
+
     } catch (err) {
         console.error(err)
     }
